requirements.txt
flask
kafka-python
eventlet
pandas
